優化每月平均天數顯示到小數點下兩位

+++ ShiftManagementSystem/Models/DTOs/EmployeeYearlyStatsDto.cs
@@ -8,7 +8,7 @@
 		public int UserId { get; set; }
 		public string FullName { get; set; }
 		public int TotalYearlyShifts { get; set; }
-		public int AverageMonthlyShifts { get; set; }
+		public decimal AverageMonthlyShifts { get; set; }
 		public Dictionary<int, int> MonthlyBreakdown { get; set; } = new();  // 月份 -> 天數
 		public decimal AverageCompliancePercentage { get; set; }
 	}
